Make activity indicator as a package
