+++ apps/fluent-tester/package.json
@@ -27,6 +27,7 @@
     "@fluentui-react-native/android-theme": ">=0.3.0 <1.0.0",
     "@fluentui-react-native/apple-theme": ">=0.5.0 <1.0.0",
     "@fluentui-react-native/default-theme": ">=0.6.0 <1.0.0",
+    "@fluentui-react-native/experimental-activity-indicator": ">=0.1.0 <1.0.0",
     "@fluentui-react-native/experimental-avatar": ">=0.9.0 <1.0.0",
     "@fluentui-react-native/experimental-button": "0.6.0",
     "@fluentui-react-native/experimental-native-button": ">=0.6.0 <1.0.0",
